Resolve merge conflicts by combining both sides:

- Borealis migration (main): CLI/data/log/sentry/discord/update/file-select
  now provided by borealis libraries; removed old implementations
  (file_select, crash_reporting, crash_handler, update_check, http, etc.)
- Keep cn-enhancements features: TPHD layer (ENABLE_TPHD switch,
  CLEAR_TPHD_FOLDER, 512MB mem1, set_hd_content_path), portable mode
  (--portable / portable.txt via borealis userDirectoryOverride),
  i18n tokenized UI text
- GAME_LIBS: borealis set + ZLIB::ZLIB (TphdPack needs zlib)
- .gitmodules: aurora -> encounter/aurora, add extern/borealis (cloned)
- aurora updated to 8005d336; adapt aurora-rmlui-translate-callback.patch
  to new baseline (gpu.cpp g_vsyncEnabled, rmlui.hpp set_glass_light_dir)
- Android: borealis-application.gradle integration; signing moved to CI
- Replace aurora::Module with borealis::Log (compatible interface)
- Add fs_path_to_string back to dusk::io (used by tphd/i18n)
- Add AIMING_RETICLE i18n tokens to all four language files
- Update AGENTS.md to reflect post-merge state

+++ src/dusk/stubs.cpp
@@ -908,7 +908,7 @@ void AIInit(u8* stack) {
     // In a real scenario, it would set up the audio interface and prepare it for use.
 }
 
-void AIInitDMA(u32 start_addr, u32 length) {
+void AIInitDMA(uintptr_t start_addr, u32 length) {
     STUB_LOG();
 }
